refactor(am): extract tool pipeline orchestration
